line_bot/requirements.txt
Flask==0.12.2
line-bot-sdk==1.5.0
pandas == 1.0.3
psycopg2
beautifulsoup4==4.9.0
requests